Move load code into dataloader
