+++ typedload/plugins/attrload.py
@@ -1,6 +1,5 @@
 """
-typedload
-Module to load data into data structures from the "attr" module
+DEPRECATED: This does nothing
 """
 
 # Copyright (C) 2018-2019 Salvo "LtWorf" Tomaselli
@@ -21,80 +20,8 @@ Module to load data into data structures from the "attr" module
 # author Salvo "LtWorf" Tomaselli <tiposchi@tiscali.it>
 
 
-from ..dataloader import _namedtupleload
-from ..exceptions import *
-from ..typechecks import is_attrs
-
-class _FakeNamedTuple(tuple):
-    """
-    This class simulates a Python3.6 NamedTuple
-    instance.
-
-    It has the same hidden fields, so the same
-    loader for the NamedTuple.
-
-    It needs to be created with fields, field_types, field_defaults
-    """
-
-    def __new__(cls, fields):
-        return super(_FakeNamedTuple, cls).__new__(cls, tuple(fields))
-
-    @property
-    def _fields(self):
-        return self[0]
-
-    @property
-    def __annotations__(self):
-        return self[1]
-
-    @property
-    def _field_defaults(self):
-        return self[2]
-
-    def __call__(self, **kwargs):
-        try:
-            return self[3](**kwargs)
-        except TypeError as e:
-            raise TypedloadTypeError(str(e), type_=self[3], value=kwargs)
-        except Exception as e:
-            raise TypedloadException(str(e), type_=self[3], value=kwargs)
-
-
-def _attrload(l, value, type_):
-    if not isinstance(value, dict):
-        raise TypedloadTypeError('Expected dictionary, got %s' % type(value), type_=type_, value=value)
-    value = value.copy()
-    names = []
-    defaults = {}
-    types = {}
-
-    for attribute in type_.__attrs_attrs__:
-        names.append(attribute.name)
-        types[attribute.name] = attribute.type
-        defaults[attribute.name] = attribute.default
-
-        # Manage name mangling
-        if 'name' in attribute.metadata:
-            dataname = attribute.metadata['name']
-            pyname = attribute.name
-
-            if dataname in value:
-                tmp = value[dataname]
-                del value[dataname]
-                value[pyname] = tmp
-
-    t = _FakeNamedTuple((
-        tuple(names),
-        types,
-        defaults,
-        type_,
-    ))
-
-    return _namedtupleload(l, value, t)
-
-
 def add2loader(l) -> None:
     """
-    Adds the "attr" handler to an existing loader
+    DEPRECATED: Calls to this can be safely removed.
     """
-    l.handlers.append((is_attrs, _attrload))
+    pass
